# Add vectorized rand! for uniform distribution