--- a/src/univariate/continuous/uniform.jl
+++ b/src/univariate/continuous/uniform.jl
@@ -111,6 +111,9 @@ end
 
 rand(rng::AbstractRNG, d::Uniform) = d.a + (d.b - d.a) * rand(rng)
 
+rand!(rng::AbstractRNG, d::Uniform, A::AbstractArray) =
+    A .= quantile.(d, rand!(rng, A))
+
 
 #### Fitting