@Thalassocracy 2021-02-07
Bonus—Vaguely improved compatibility and performance with CUDA.jl:

```
using CUDA
using Random
using Distributions

CUDA.allowscalar(false)

t = CuArray{Float64}(undef, 1000)

rand!(t) # works

rand!(CURAND.default_rng(), Uniform(0,1), t) # doesn't work currently
```

This MWE doesn't work on current Distributions.jl but will work in this pull request, although the reason for the error is more of a CUDA.jl problem than a Distributions.jl problem. 

`rand!(Uniform(0,1), CuArray{Float64}(undef, 1000))` will work when `CUDA.allowscalar(true)` under both this pull request and the current master branch of Distributions.jl, but it requires a lot of scalar indexing and is not performant.